Support seeking.
This allows ffmpeg to seek in the output file. This is kinda a hack, as
we still stream the file so we will be serving "wrong" data when live
transcoding, however for matroska (our only currently used output
container) the only thing that needs to be updated retrospectively is
the location of the seek cues. This is not needed for streaming the file
(this is what we were doing before anyways) however once the transcode
is complete it can be updated afterwards so that future viewers running
on the cached file can have the seek information available and enable
seeking.

Also switches the code to use a std::file::File rather then a RawFd and
doesn't leak descriptors anymore. (Well except we never clear the cache
anyways so they actually stick around.)